+++ src/middleware/file-extension.js
@@ -1,9 +1,6 @@
 'use strict';
 
 module.exports = {
-  // file extension BalckList
-  extensionBlacklist: new Set(['', '.bz2', '.dcm', '.exe', '.gz', '.rar', '.iso', '.zip']),
-
   // file extension WhiteList
   extensionWhitelist: new Set([
     '.webp',