+++ src/middleware/request-id.js
@@ -1,26 +1,13 @@
 'use strict';
 
-const req = require('express/lib/request');
 const uuid = require('uuid');
 
-function getRequestId(options) {
-  options = options || {};
+function getRequestId() {
   const uuidVersion = 'v4';
-  const headerName = 'X-Request-Id';
-  const RequestId = req.header[headerName.toLowerCase()] || uuid[uuidVersion](options, options.buffer, options.offset);
-  return RequestId;
+  const Request_Id = uuid[uuidVersion]();
+  return Request_Id;
 }
 
-// set headers respone
-function setHeadersRequestId() {
-  const headers = { 'X-Request-Id': getRequestId() };
-  return headers;
-}
-
-// Sent RequestId to headers reespone
-const RequestId = setHeadersRequestId();
-
-// Logging
-// TODO Create logging for trace error client
+const RequestId = getRequestId();
 
 module.exports = { RequestId };