+++ test/index.js
@@ -1,23 +1,65 @@
 'use strict';
 
-const req = require('express/lib/request');
+// const req = require('express/lib/request');
+// const uuid = require('uuid');
+// const https = require('https');
+
+// module.exports = {
+//   RequestId: { 'X-Request-Id': id },
+//   RequestId_Log: id
+// }
+
+// /**
+// Handles an incoming response from an upstream server, piping it to the client
+// or handling errors as appropriate.
+
+// @param {https.ClientRequest} req
+//   Express request object.
+
+// @param {https.ServerResponse} res
+//   Express response object.
+// **/
+
+// function getRequestId(req) {
+//   const uuidVersion = 'v4';
+//   const headerName = 'X-Request-Id';
+//   const RequestId = req.headers[headerName.toLowerCase()] || uuid[uuidVersion]();
+//   return RequestId;
+// }
+
+// const id = getRequestId();
+
+// Sent RequestId to headers reespone
+// const RequestId = { 'X-Request-Id': id };
+
+// Logging
+// const RequestId_Log = id;
+
+// console.log(id);
+
+// module.exports = { RequestId, RequestId_Log };
+
 const uuid = require('uuid');
 
-function getRequestId(options) {
-  options = options || {};
+function getRequestId(req, res, next) {
+  // options = options || {};
   const uuidVersion = 'v4';
   const headerName = 'X-Request-Id';
-  const RequestId = req.header[headerName.toLowerCase()] || uuid[uuidVersion](options, options.buffer, options.offset);
-  return RequestId;
+  const Request_Id = req.headers[headerName.toLowerCase()] || uuid[uuidVersion]();
+  return Request_Id;
+
+  // return function (req, res, next) {
+  //   const Request_Id = req.headers[headerName.toLowerCase()] || uuid[uuidVersion]();
+  //   return Request_Id;
+  // };
 }
 
-// set headers respone
-function setHeadersRequestId() {
-  const headers = { 'X-NCD-Request-Id': getRequestId() };
-  return headers;
-}
+const id = getRequestId();
 
-// Kirim headers ke headersRespone.js
-const RequestId = setHeadersRequestId();
+// Sent RequestId to headers reespone
+const RequestId = { 'X-Request-Id': id };
 
-console.log(RequestId);
+// Logging
+const RequestId_Log = id;
+
+console.log(id);